feat: add app duration to registry schema and expiration tracking (#3307)
### Description

Added app expiration functionality to the app registry and account systems. Apps now have a configurable duration that determines how long they remain active after installation, with a maximum limit of 365 days.

### Changes

- Added `duration` field to the App struct in the registry schema
- Added validation for app duration with a maximum of 365 days
- Implemented app expiration tracking in the account system
- Added `getAppExpiration` method to retrieve when an app will expire
- Updated `isAppEntitled` to check if an app is still within its valid period
- Removed manifest verification that's no longer needed
- Fixed permission checks to properly verify client access

### Checklist

- [x] Tests added where required
- [x] Documentation updated where applicable
- [x] Changes adhere to the repository's contribution guidelines

+++ packages/contracts/src/apps/ITownsApp.sol
@@ -24,5 +24,5 @@ interface ITownsApp is IERC6900Module, IERC6900ExecutionModule {
 
     /// @notice Returns the access duration of the app
     /// @return accessDuration The access duration of the app
-    function accessDuration() external view returns (uint64);
+    function accessDuration() external view returns (uint48);
 }
